center points can be floats

+++ priv/repo/migrations/20161116024603_create_obstacle.exs
@@ -7,8 +7,8 @@ defmodule Bolide.Repo.Migrations.CreateObstacle do
       add :section_id, references(:sections, on_delete: :nothing)
       # what I really want
       # add :circle, :circle
-      add :center_x, :integer
-      add :center_y, :integer
+      add :center_x, :float
+      add :center_y, :float
       add :radius, :float
 
       timestamps()
